Issue #377 - use the VMUSB drivers in place rather than as copies
Use the delay driver directly from usb/vmusb to improve maintainability.
